Error created while trying to learn.
I'm trying to learn Sonata and Symfony. I called the entity CategoryEntity. The following service entry worked. While trying to troubleshoot an error when clicking to add a category I changed CategoryEntity to Category (as in the unmodified sample). That created the error following the service entry. Putting the service entry back will not fix the error.

    admin.category:
      class: App\Admin\CategoryAdmin
      arguments: [~, App\Entity\CategoryEntity, ~]
      tags:
        - { name: sonata.admin, manager_type: orm, label: Category }

An exception has been thrown during the rendering of a template ("Unable to generate a URL for the named route "admin_app_categoryentity_create" as such route does not exist.").